requirements.txt
PyQt6==6.6.1
psycopg2-binary
python-dotenv==1.0.1